move file related fn to file module
