+++ src/lib.rs
@@ -1,4 +1,6 @@
 pub mod cgroups;
+pub mod error;
+pub mod file_manager;
 pub mod handler;
 pub mod model;
 mod tests;
